finsight-ai/backend/requirements.txt
# ============================================================
# FinSight AI — Backend Dependencies
# ============================================================

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Google Gemini
google-generativeai==0.8.4

# Market data
yfinance==0.2.51
nsepython==2.8

# Technical analysis
pandas-ta==0.3.14b1
pandas==2.2.3
numpy==2.2.1

# Vector DB (RAG)
chromadb==0.6.3

# Web scraping
beautifulsoup4==4.13.3
requests==2.32.3
lxml==5.3.1

# Reddit API
praw==7.8.1

# RSS feed parsing
feedparser==6.0.11

# Scheduler
apscheduler==3.11.0

# Database
supabase==2.13.0

# Environment variables
python-dotenv==1.0.1

# Async HTTP
httpx==0.28.1
aiohttp==3.11.12

# Logging
loguru==0.7.3

# CORS
starlette==0.45.3
